front-end refactor
